+++ packages/driver-prettier/CHANGELOG.md
@@ -3,3 +3,4 @@
 #### 🎉 Release
 
 - Initial release!
+- Added a `prettier.create-ignore-file` event.